se actualizo el codigo y se le hizo cambios

+++ Programa/Equipaje.java
@@ -1,9 +1,9 @@
 public class Equipaje {
     int id;
-    double peso;
+    int peso;
     int idPasajero;
 
-    public Equipaje(int id, double peso, int idPasajero) {
+    public Equipaje(int id, int peso, int idPasajero) {
         this.id = id;
         this.peso = peso;
         this.idPasajero = idPasajero;
